+++ rust_snuba/Cargo.toml
@@ -4,6 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[[bin]]
+name = "querylog_consumer"
+path = "src/bin/consumer/querylog_consumer.rs"
 
 [dependencies]
 chrono = "0.4.11"
