+++ rust_snuba/src/bin/consumer/other/errors.rs
