+++ rust_snuba/src/bin/consumer/other/noop.rs
