+++ rust_snuba/src/consumer/mod.rs
@@ -1,0 +1,1 @@
+pub mod querylog_consumer;
