+++ rust_snuba/src/lib.rs
@@ -2,6 +2,7 @@ pub mod backends;
 pub mod processing;
 pub mod types;
 pub mod utils;
+pub mod consumer;
 
 #[cfg(test)]
 mod tests {
